Problem: asset DB errors are not propagated to REST API
Solution: add error propagating solution based on utils_web and _error
array. Basically all functions are supposed to return integer and
formatted string with error message. Scheme is compatible with all the
DB functions we have.

Then support macros exists to throw it upper or to print it from REST
API correctly.

There is no parallel structure of internal errors, which needs to be
mapped to REST API.

Signed-off-by: Michal Vyskocil <michalvyskocil@eaton.com>